Fixed test failure
Was caused I think by moving the rsdb open to be in the startup code
thus slowing down the start method for core, this caused the gossiping
of the manually added peers (done in the test after starting each
previously stopped node )to happen after the validator peer handler
had removed all the peers resulting in gossiping certs to no one. Thus
leaving the network in a disconnected state.

The fix was really hacky, but just to wait 250ms after starting a node
and hoping that the validator peer handler will have removed nodes by
that point and then add the peers and then wait a little for them to be
registered and then finally gossip to them.